Enabled the FPU

+++ inc/Platform/Parameters/PlatformParameters.hpp
@@ -35,7 +35,7 @@ namespace PlatformParameters {
     // OBDH parameters
     inline Parameter<uint16_t> obcBootCounter(0);
     inline Parameter<uint64_t> obcSysTick(0);
-    inline Parameter<int8_t> mcuTemperature(0);
+    inline Parameter<float> mcuTemperature(0);
 }
 
 #endif
